Fix NCURSES_WIDECHAR preprocessor test
A build from master failed for me because the type wint_t wasn't found.
This type was referenced on line 1743, in an #if block predicated on the
condition `NCURSES_WIDECHAR > 0`. The only time the variables defined in
this block are referenced are also inside of an #if block, this time
predicated by the additional condition `NCURSES_WIDECHAR > 0 && defined
HAVE_NCURSESW`. It seems to be an error that this extra condition was
omitted from the first preprocessor block.